chore: remove shift+space mapping from f265d72
Remove the mapping for Shift + Space, so it's handled as usual: space.

f265d72 introduced mappings for a ton of key combinations, to allow one
to bind combinations usually not handled by terminal emulators. This
commit removes the one for Shift + Space, as that won't ever be used by
me for key bindings on terminal programs.